requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pandas
numpy
scikit-learn
python-jose
passlib[bcrypt]
twilio
python-multipart